requirements.txt
flask
sentence-transformers
faiss-cpu
torch
numpy
pandas
